test/requirements.txt
pika
paho-mqtt
pika
pandas
requests
matplotlib
prophet
plotly
paho-mqtt